Update and rename DataStructure.py to LinkedListAVLDictionaryHybrid.py
optimizing modularity, adding comments, fixing stuff, half the work on the class
